Commit final do projeto

+++ src/br/hue/hue/inf008/kinkinmonitor/controller/AreaMonitoradaController.java
@@ -1,8 +1,9 @@
 package br.hue.hue.inf008.kinkinmonitor.controller;
 
+import java.util.List;
+
 import br.hue.hue.inf008.kinkinmonitor.model.AreaMonitorada;
 import br.hue.hue.inf008.kinkinmonitor.persistence.AreaMonitoradaDAO;
-import java.util.List;
 
 public class AreaMonitoradaController implements IController<AreaMonitorada> {
 
@@ -20,7 +21,7 @@ public class AreaMonitoradaController implements IController<AreaMonitorada> {
 	@Override
 	public AreaMonitorada findById(String nome) throws Exception {
 		if (nome == null || nome.isEmpty()) {
-			throw new Exception("Informe o Identificador da Área Monitorada.");
+			throw new Exception("Identificador deve ser preenchido.");
 		}
 		return this.dao.findByNome(nome);
 	}
@@ -28,7 +29,7 @@ public class AreaMonitoradaController implements IController<AreaMonitorada> {
 	@Override
 	public int insert(AreaMonitorada domain) throws Exception {
 		if (domain.getNome() == null || domain.getNome().isEmpty()) {
-			throw new Exception("Informe o Identificador da Área Monitorada.");
+			throw new Exception("Identificador deve ser preenchido.");
 		}
 		return this.dao.insert(domain);
 	}
@@ -36,13 +37,16 @@ public class AreaMonitoradaController implements IController<AreaMonitorada> {
 	@Override
 	public int update(AreaMonitorada domain) throws Exception {
 		if (domain.getNome() == null || domain.getNome().isEmpty()) {
-			throw new Exception("Informe o Identificador da Área Monitorada.");
+			throw new Exception("Identificador deve ser preenchido.");
 		}
 		return this.dao.update(domain);
 	}
 
 	@Override
 	public int delete(AreaMonitorada domain) throws Exception {
+		if (domain == null || domain.getId() == 0) {
+			throw new Exception("Área Monitorada deve ser informada.");
+		}
 		return this.dao.delete(domain);
 	}
